+++ lib/cachecontrol/adapter.py
@@ -1,7 +1,8 @@
-from requests.adapters import HTTPAdapter
+from lib.requests.adapters import HTTPAdapter
 
 from .controller import CacheController
 from .cache import DictCache
+
 
 class CacheControlAdapter(HTTPAdapter):
     invalidating_methods = set(['PUT', 'DELETE'])
@@ -58,7 +59,11 @@ class CacheControlAdapter(HTTPAdapter):
                 response = cached_response
             else:
                 # try to cache the response
-                self.controller.cache_response(request, response)
+                try:
+                    self.controller.cache_response(request, response)
+                except Exception as e:
+                    # Failed to cache the results
+                    pass
 
         resp = super(CacheControlAdapter, self).build_response(
             request, response